# openapi.yaml
openapi: 3.0.0
info:
  title: BOGOWI Blockchain API
  description: High-performance Go API for BOGOWI blockchain operations including tokens, NFTs, rewards, and DAO functionality
  version: 1.0.0
  contact:
    name: BOGOWI Team
    url: https://bogowi.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.bogowi.com/api
    description: Production server
paths:
...
  /token/balance/{address}:
    get:
      summary: Get Token Balance
      description: Returns the BOGO token balance for a specific address
      tags: [Tokens]
      parameters:
        - name: address
          in: path
          required: true
          description: Ethereum wallet address
          schema:
            type: string
            example: "0x742b18C3E6C2E0dD5f75FbBd7D71d8CaE59c7054"
      responses:
        '200':
          description: Token balance retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  balance:
                    type: string
                    example: "1000.5"
        '400':
          description: Invalid address format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /token/flavored-balances/{address}:
    get:
      summary: Get Flavored Token Balances
      description: Returns balances for Ocean, Earth, and Wildlife BOGO tokens
      tags: [Tokens]
      parameters:
        - name: address
          in: path
          required: true
          description: Ethereum wallet address
          schema:
            type: string
      responses:
        '200':
          description: Flavored token balances
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  balances:
                    type: object
                    properties:
                      ocean:
                        type: string
                      earth:
                        type: string
                      wildlife:
                        type: string
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
...